pyproject.toml
[tool.poetry]
name = "anna-lee-system"
version = "0.1.0"
description = "Data Science/ Machine Learning practice on the Instagram art platform"
authors = ["devwonyouu <jangwonyouu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
instapy = "^0.6.16"
python-dotenv = "^0.20.0"
sqlmodel = "^0.0.6"
fastapi = "^0.77.1"
click = "^8.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
